RSC convert admin patient-acquisition page
